Remove some dependencies in System.IO in the common library.
