+++ DK.Common.Tests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NUnit" version="3.13.3" targetFramework="net472" />
+  <package id="NUnit3TestAdapter" version="4.2.1" targetFramework="net472" />
+</packages>
