+++ DK.Common/DK.Common.csproj
@@ -39,10 +39,6 @@
     <AssemblyOriginatorKeyFile>StrongNameKey.pfx</AssemblyOriginatorKeyFile>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Microsoft.VisualStudio.Setup.Configuration.Interop, Version=1.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.VisualStudio.Setup.Configuration.Interop.2.3.2262-g94fae01e\lib\net35\Microsoft.VisualStudio.Setup.Configuration.Interop.dll</HintPath>
-      <EmbedInteropTypes>False</EmbedInteropTypes>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -56,11 +52,9 @@
     <Compile Include="AppEnvironment\BuiltInFileSystem.cs" />
     <Compile Include="AppEnvironment\DkAppSettings.cs" />
     <Compile Include="AppEnvironment\DkEnvironment.cs" />
-    <Compile Include="AppEnvironment\DkEnvVarMerger.cs" />
-    <Compile Include="AppEnvironment\EnvVarList.cs" />
-    <Compile Include="AppEnvironment\FileHelper.cs" />
     <Compile Include="AppEnvironment\DkAppContext.cs" />
     <Compile Include="AppEnvironment\IFileSystem.cs" />
+    <Compile Include="AppEnvironment\PathUtil.cs" />
     <Compile Include="AppEnvironment\RegistryKeyHelper.cs" />
     <Compile Include="CodeAnalysis\CAErrorMarker.cs" />
     <Compile Include="CodeAnalysis\CAErrors.cs" />
@@ -276,15 +270,7 @@
     <Compile Include="TriState.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="packages.config" />
     <None Include="StrongNameKey.pfx" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <Import Project="..\packages\Microsoft.VisualStudio.Setup.Configuration.Interop.2.3.2262-g94fae01e\build\Microsoft.VisualStudio.Setup.Configuration.Interop.targets" Condition="Exists('..\packages\Microsoft.VisualStudio.Setup.Configuration.Interop.2.3.2262-g94fae01e\build\Microsoft.VisualStudio.Setup.Configuration.Interop.targets')" />
-  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
-    <PropertyGroup>
-      <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
-    </PropertyGroup>
-    <Error Condition="!Exists('..\packages\Microsoft.VisualStudio.Setup.Configuration.Interop.2.3.2262-g94fae01e\build\Microsoft.VisualStudio.Setup.Configuration.Interop.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\Microsoft.VisualStudio.Setup.Configuration.Interop.2.3.2262-g94fae01e\build\Microsoft.VisualStudio.Setup.Configuration.Interop.targets'))" />
-  </Target>
 </Project>
