--- DK.Common/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.VisualStudio.Setup.Configuration.Interop" version="2.3.2262-g94fae01e" targetFramework="net472" />
-</packages>
