+++ DkTools/packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Microsoft.VisualStudio.SDK.Analyzers" version="15.8.33" targetFramework="net461" />
+  <package id="Microsoft.VisualStudio.Setup.Configuration.Interop" version="3.1.2196" targetFramework="net472" />
   <package id="Microsoft.VisualStudio.Threading.Analyzers" version="15.8.122" targetFramework="net461" />
   <package id="Microsoft.VSSDK.BuildTools" version="15.8.3252" targetFramework="net461" developmentDependency="true" />
   <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
